refactor: align apollo on relay, extract common RepoItem component
