+++ apollo/RepoList.tsx
@@ -2,34 +2,32 @@ import { useState } from 'react';
 import type { TypedDocumentNode } from '@apollo/client';
 import { gql } from '@apollo/client';
 import { useQuery } from '@apollo/client/react';
-import { Heading, Icon, Link, ListItem, Text, Badge } from '@chakra-ui/react';
-import { faCodePullRequest } from '@fortawesome/free-solid-svg-icons';
-import { FontAwesomeIcon } from '@fortawesome/react-fontawesome';
-
-import type { GetRepositoriesQuery, GetRepositoriesQueryVariables } from '../app/apollo/__gen__/graphql';
+import { Heading } from '@chakra-ui/react';
 
 import { Ul } from 'components/ChakraMdxProvider';
 import { Spinner } from 'components/Spinner';
 import InfiniteScrollTrigger from 'components/InfiniteScrollTrigger';
-import InternalLink from 'components/InternalLink';
 import { Checkbox } from 'components/ui/checkbox';
 
+import type { GetRepositoriesQuery, GetRepositoriesQueryVariables } from '../app/apollo/__gen__/graphql';
+
+import { RepoItem } from 'components/RepoItem';
+
+interface RepoListProps {
+    login: string;
+}
+
 export const REPOS_QUERY: TypedDocumentNode<GetRepositoriesQuery, GetRepositoriesQueryVariables> = gql`
-    query GetRepositories($login: String!, $after: String, $first: Int!) {
+    query GetRepositories($login: String!, $cursor: String, $count: Int!) {
         repositoryOwner(login: $login) {
             login
             id
             repositories(
-                first: $first
-                after: $after
                 orderBy: { field: NAME, direction: ASC }
+                first: $count
+                after: $cursor
                 ownerAffiliations: [OWNER]
             ) {
-                totalCount
-                pageInfo {
-                    hasNextPage
-                    endCursor
-                }
                 edges {
                     node {
                         id
@@ -43,21 +41,23 @@ export const REPOS_QUERY: TypedDocumentNode<GetRepositoriesQuery, GetRepositorie
                         }
                     }
                 }
+                pageInfo {
+                    endCursor
+                    hasNextPage
+                }
+                totalCount
             }
         }
     }
 `;
 
-interface RepoListProps {
-    login: string;
-}
-
 export default function RepoList({ login }: RepoListProps) {
     const [showAll, setShowAll] = useState(false);
+
     const { loading, error, data, fetchMore } = useQuery<GetRepositoriesQuery, GetRepositoriesQueryVariables>(
         REPOS_QUERY,
         {
-            variables: { login, first: 10 },
+            variables: { login, count: 10 },
         }
     );
 
@@ -69,16 +69,19 @@ export default function RepoList({ login }: RepoListProps) {
     if (!data) return null;
 
     const { repositoryOwner } = data;
+
     if (!repositoryOwner?.repositories) return <div>No repositories found</div>;
 
     const { repositories } = repositoryOwner;
     const { totalCount, edges, pageInfo } = repositories;
 
-    const loadMore = async (after: string) => {
+    if (totalCount == 0 || edges?.length == 0) return <div>No repositories found</div>;
+
+    const loadMore = async (cursor: string) => {
         if (!pageInfo.hasNextPage) return;
 
         await fetchMore({
-            variables: { after, login },
+            variables: { cursor, login },
         });
     };
 
@@ -105,7 +108,7 @@ export default function RepoList({ login }: RepoListProps) {
                         if (!node) {
                             return (
                                 <InfiniteScrollTrigger
-                                    key={idx}
+                                    key={edges.length - 1}
                                     enabled={isLastElement && pageInfo.hasNextPage}
                                     onLoadMore={onLoadMore}>
                                     <div />
@@ -129,44 +132,5 @@ export default function RepoList({ login }: RepoListProps) {
 
             {loading && <Spinner size="sm" />}
         </>
-    );
-}
-
-interface RepoItemProps {
-    hideIfFork?: boolean;
-    repo: {
-        name: string;
-        nameWithOwner: string;
-        isFork: boolean;
-        url: string;
-        pullRequests: {
-            totalCount: number;
-        };
-    };
-}
-
-function RepoItem({ repo, hideIfFork }: RepoItemProps) {
-    const { name, nameWithOwner, pullRequests, url, isFork } = repo;
-    const { totalCount } = pullRequests;
-
-    if (isFork && hideIfFork) return null;
-
-    return (
-        <ListItem alignItems="center" gap="1">
-            <InternalLink prefetch={false} href={'./' + nameWithOwner}>
-                {name}
-            </InternalLink>
-
-            {isFork && <Badge>fork</Badge>}
-            {totalCount > 0 && <span> - {totalCount} PRs</span>}
-            {totalCount > 0 && (
-                <Link href={url} rel="noopener noreferrer nofollow" target="_blank">
-                    - <Text> open in GitHub</Text>
-                    <Icon ml={1}>
-                        <FontAwesomeIcon icon={faCodePullRequest} />
-                    </Icon>
-                </Link>
-            )}
-        </ListItem>
     );
 }
