+++ app/apollo/__gen__/gql.ts
@@ -17,7 +17,7 @@ type Documents = {
     '\n    fragment CommitWithStatuses_commit on Commit {\n        id\n        authoredDate\n        status {\n            id\n            commit {\n                oid\n            }\n            contexts {\n                avatarUrl\n                context\n                #creator { login }\n                state\n                description\n                targetUrl\n            }\n            state\n        }\n        commitUrl\n        message\n        # would break overall loading the whole page. shrug\n        # author { user { login name avatarUrl } }\n    }\n': typeof types.CommitWithStatuses_CommitFragmentDoc;
     '\n    fragment BranchInfoRowFragment_ref on Ref {\n        name\n        id\n        target {\n            ... on Commit {\n                id\n                ...CommitWithStatuses_commit\n            }\n        }\n        associatedPullRequests(first: 1, states: [OPEN]) {\n            edges {\n                node {\n                    id\n                    ...PullRequestMergeFragment_ref\n                }\n            }\n        }\n    }\n    \n    \n': typeof types.BranchInfoRowFragment_RefFragmentDoc;
     '\n    fragment PullRequestMergeFragment_ref on PullRequest {\n        id\n        headRefOid\n        number\n        url\n        title\n        mergeStateStatus\n        closed\n        isDraft\n        merged\n        mergeable\n    }\n': typeof types.PullRequestMergeFragment_RefFragmentDoc;
-    '\n    query GetRepositories($login: String!, $after: String, $first: Int!) {\n        repositoryOwner(login: $login) {\n            login\n            id\n            repositories(\n                first: $first\n                after: $after\n                orderBy: { field: NAME, direction: ASC }\n                ownerAffiliations: [OWNER]\n            ) {\n                totalCount\n                pageInfo {\n                    hasNextPage\n                    endCursor\n                }\n                edges {\n                    node {\n                        id\n                        name\n                        nameWithOwner\n                        isFork\n                        url\n                        description\n                        pullRequests(first: 1, states: [OPEN]) {\n                            totalCount\n                        }\n                    }\n                }\n            }\n        }\n    }\n': typeof types.GetRepositoriesDocument;
+    '\n    query GetRepositories($login: String!, $cursor: String, $count: Int!) {\n        repositoryOwner(login: $login) {\n            login\n            id\n            repositories(\n                orderBy: { field: NAME, direction: ASC }\n                first: $count\n                after: $cursor\n                ownerAffiliations: [OWNER]\n            ) {\n                edges {\n                    node {\n                        id\n                        name\n                        nameWithOwner\n                        isFork\n                        url\n                        description\n                        pullRequests(first: 1, states: [OPEN]) {\n                            totalCount\n                        }\n                    }\n                }\n                pageInfo {\n                    endCursor\n                    hasNextPage\n                }\n                totalCount\n            }\n        }\n    }\n': typeof types.GetRepositoriesDocument;
     '\n    query UserFragmentStoryQuery {\n        repositoryOwner(login: "test-id") {\n            ... on RepositoryOwner {\n                ...UserFragment_repositoryOwner\n            }\n        }\n    }\n    \n': typeof types.UserFragmentStoryQueryDocument;
     '\n    fragment UserFragment_repositoryOwner on RepositoryOwner {\n        ... on User {\n            id\n            login\n            company\n            avatarUrl\n        }\n        ... on Organization {\n            id\n            login\n            avatarUrl\n        }\n    }\n': typeof types.UserFragment_RepositoryOwnerFragmentDoc;
     '\n    mutation useMergePRMutation($input: MergePullRequestInput!) {\n        mergePullRequest(input: $input) {\n            pullRequest {\n                merged\n                state\n                mergeStateStatus\n            }\n        }\n    }\n': typeof types.UseMergePrMutationDocument;
@@ -33,7 +33,7 @@ const documents: Documents = {
         types.BranchInfoRowFragment_RefFragmentDoc,
     '\n    fragment PullRequestMergeFragment_ref on PullRequest {\n        id\n        headRefOid\n        number\n        url\n        title\n        mergeStateStatus\n        closed\n        isDraft\n        merged\n        mergeable\n    }\n':
         types.PullRequestMergeFragment_RefFragmentDoc,
-    '\n    query GetRepositories($login: String!, $after: String, $first: Int!) {\n        repositoryOwner(login: $login) {\n            login\n            id\n            repositories(\n                first: $first\n                after: $after\n                orderBy: { field: NAME, direction: ASC }\n                ownerAffiliations: [OWNER]\n            ) {\n                totalCount\n                pageInfo {\n                    hasNextPage\n                    endCursor\n                }\n                edges {\n                    node {\n                        id\n                        name\n                        nameWithOwner\n                        isFork\n                        url\n                        description\n                        pullRequests(first: 1, states: [OPEN]) {\n                            totalCount\n                        }\n                    }\n                }\n            }\n        }\n    }\n':
+    '\n    query GetRepositories($login: String!, $cursor: String, $count: Int!) {\n        repositoryOwner(login: $login) {\n            login\n            id\n            repositories(\n                orderBy: { field: NAME, direction: ASC }\n                first: $count\n                after: $cursor\n                ownerAffiliations: [OWNER]\n            ) {\n                edges {\n                    node {\n                        id\n                        name\n                        nameWithOwner\n                        isFork\n                        url\n                        description\n                        pullRequests(first: 1, states: [OPEN]) {\n                            totalCount\n                        }\n                    }\n                }\n                pageInfo {\n                    endCursor\n                    hasNextPage\n                }\n                totalCount\n            }\n        }\n    }\n':
         types.GetRepositoriesDocument,
     '\n    query UserFragmentStoryQuery {\n        repositoryOwner(login: "test-id") {\n            ... on RepositoryOwner {\n                ...UserFragment_repositoryOwner\n            }\n        }\n    }\n    \n':
         types.UserFragmentStoryQueryDocument,
@@ -87,8 +87,8 @@ export function graphql(
  * The graphql function is used to parse GraphQL queries into a document that can be used by GraphQL clients.
  */
 export function graphql(
-    source: '\n    query GetRepositories($login: String!, $after: String, $first: Int!) {\n        repositoryOwner(login: $login) {\n            login\n            id\n            repositories(\n                first: $first\n                after: $after\n                orderBy: { field: NAME, direction: ASC }\n                ownerAffiliations: [OWNER]\n            ) {\n                totalCount\n                pageInfo {\n                    hasNextPage\n                    endCursor\n                }\n                edges {\n                    node {\n                        id\n                        name\n                        nameWithOwner\n                        isFork\n                        url\n                        description\n                        pullRequests(first: 1, states: [OPEN]) {\n                            totalCount\n                        }\n                    }\n                }\n            }\n        }\n    }\n'
-): (typeof documents)['\n    query GetRepositories($login: String!, $after: String, $first: Int!) {\n        repositoryOwner(login: $login) {\n            login\n            id\n            repositories(\n                first: $first\n                after: $after\n                orderBy: { field: NAME, direction: ASC }\n                ownerAffiliations: [OWNER]\n            ) {\n                totalCount\n                pageInfo {\n                    hasNextPage\n                    endCursor\n                }\n                edges {\n                    node {\n                        id\n                        name\n                        nameWithOwner\n                        isFork\n                        url\n                        description\n                        pullRequests(first: 1, states: [OPEN]) {\n                            totalCount\n                        }\n                    }\n                }\n            }\n        }\n    }\n'];
+    source: '\n    query GetRepositories($login: String!, $cursor: String, $count: Int!) {\n        repositoryOwner(login: $login) {\n            login\n            id\n            repositories(\n                orderBy: { field: NAME, direction: ASC }\n                first: $count\n                after: $cursor\n                ownerAffiliations: [OWNER]\n            ) {\n                edges {\n                    node {\n                        id\n                        name\n                        nameWithOwner\n                        isFork\n                        url\n                        description\n                        pullRequests(first: 1, states: [OPEN]) {\n                            totalCount\n                        }\n                    }\n                }\n                pageInfo {\n                    endCursor\n                    hasNextPage\n                }\n                totalCount\n            }\n        }\n    }\n'
+): (typeof documents)['\n    query GetRepositories($login: String!, $cursor: String, $count: Int!) {\n        repositoryOwner(login: $login) {\n            login\n            id\n            repositories(\n                orderBy: { field: NAME, direction: ASC }\n                first: $count\n                after: $cursor\n                ownerAffiliations: [OWNER]\n            ) {\n                edges {\n                    node {\n                        id\n                        name\n                        nameWithOwner\n                        isFork\n                        url\n                        description\n                        pullRequests(first: 1, states: [OPEN]) {\n                            totalCount\n                        }\n                    }\n                }\n                pageInfo {\n                    endCursor\n                    hasNextPage\n                }\n                totalCount\n            }\n        }\n    }\n'];
 /**
  * The graphql function is used to parse GraphQL queries into a document that can be used by GraphQL clients.
  */
