+++ app/apollo/__gen__/graphql.ts
@@ -144,8 +144,8 @@ export type PullRequestMergeFragment_RefFragment = {
 
 export type GetRepositoriesQueryVariables = Exact<{
     login: string;
-    after?: string | null | undefined;
-    first: number;
+    cursor?: string | null | undefined;
+    count: number;
 }>;
 
 export type GetRepositoriesQuery = {
@@ -157,7 +157,6 @@ export type GetRepositoriesQuery = {
               repositories: {
                   __typename: 'RepositoryConnection';
                   totalCount: number;
-                  pageInfo: { __typename: 'PageInfo'; hasNextPage: boolean; endCursor: string | null };
                   edges: Array<{
                       __typename: 'RepositoryEdge';
                       node: {
@@ -171,6 +170,7 @@ export type GetRepositoriesQuery = {
                           pullRequests: { __typename: 'PullRequestConnection'; totalCount: number };
                       } | null;
                   } | null> | null;
+                  pageInfo: { __typename: 'PageInfo'; endCursor: string | null; hasNextPage: boolean };
               };
           }
         | {
@@ -180,7 +180,6 @@ export type GetRepositoriesQuery = {
               repositories: {
                   __typename: 'RepositoryConnection';
                   totalCount: number;
-                  pageInfo: { __typename: 'PageInfo'; hasNextPage: boolean; endCursor: string | null };
                   edges: Array<{
                       __typename: 'RepositoryEdge';
                       node: {
@@ -194,6 +193,7 @@ export type GetRepositoriesQuery = {
                           pullRequests: { __typename: 'PullRequestConnection'; totalCount: number };
                       } | null;
                   } | null> | null;
+                  pageInfo: { __typename: 'PageInfo'; endCursor: string | null; hasNextPage: boolean };
               };
           }
         | null;
@@ -688,12 +688,12 @@ export const GetRepositoriesDocument = {
                 },
                 {
                     kind: 'VariableDefinition',
-                    variable: { kind: 'Variable', name: { kind: 'Name', value: 'after' } },
+                    variable: { kind: 'Variable', name: { kind: 'Name', value: 'cursor' } },
                     type: { kind: 'NamedType', name: { kind: 'Name', value: 'String' } },
                 },
                 {
                     kind: 'VariableDefinition',
-                    variable: { kind: 'Variable', name: { kind: 'Name', value: 'first' } },
+                    variable: { kind: 'Variable', name: { kind: 'Name', value: 'count' } },
                     type: { kind: 'NonNullType', type: { kind: 'NamedType', name: { kind: 'Name', value: 'Int' } } },
                 },
             ],
@@ -721,16 +721,6 @@ export const GetRepositoriesDocument = {
                                     arguments: [
                                         {
                                             kind: 'Argument',
-                                            name: { kind: 'Name', value: 'first' },
-                                            value: { kind: 'Variable', name: { kind: 'Name', value: 'first' } },
-                                        },
-                                        {
-                                            kind: 'Argument',
-                                            name: { kind: 'Name', value: 'after' },
-                                            value: { kind: 'Variable', name: { kind: 'Name', value: 'after' } },
-                                        },
-                                        {
-                                            kind: 'Argument',
                                             name: { kind: 'Name', value: 'orderBy' },
                                             value: {
                                                 kind: 'ObjectValue',
@@ -750,6 +740,16 @@ export const GetRepositoriesDocument = {
                                         },
                                         {
                                             kind: 'Argument',
+                                            name: { kind: 'Name', value: 'first' },
+                                            value: { kind: 'Variable', name: { kind: 'Name', value: 'count' } },
+                                        },
+                                        {
+                                            kind: 'Argument',
+                                            name: { kind: 'Name', value: 'after' },
+                                            value: { kind: 'Variable', name: { kind: 'Name', value: 'cursor' } },
+                                        },
+                                        {
+                                            kind: 'Argument',
                                             name: { kind: 'Name', value: 'ownerAffiliations' },
                                             value: {
                                                 kind: 'ListValue',
@@ -760,18 +760,6 @@ export const GetRepositoriesDocument = {
                                     selectionSet: {
                                         kind: 'SelectionSet',
                                         selections: [
-                                            { kind: 'Field', name: { kind: 'Name', value: 'totalCount' } },
-                                            {
-                                                kind: 'Field',
-                                                name: { kind: 'Name', value: 'pageInfo' },
-                                                selectionSet: {
-                                                    kind: 'SelectionSet',
-                                                    selections: [
-                                                        { kind: 'Field', name: { kind: 'Name', value: 'hasNextPage' } },
-                                                        { kind: 'Field', name: { kind: 'Name', value: 'endCursor' } },
-                                                    ],
-                                                },
-                                            },
                                             {
                                                 kind: 'Field',
                                                 name: { kind: 'Name', value: 'edges' },
@@ -850,6 +838,18 @@ export const GetRepositoriesDocument = {
                                                     ],
                                                 },
                                             },
+                                            {
+                                                kind: 'Field',
+                                                name: { kind: 'Name', value: 'pageInfo' },
+                                                selectionSet: {
+                                                    kind: 'SelectionSet',
+                                                    selections: [
+                                                        { kind: 'Field', name: { kind: 'Name', value: 'endCursor' } },
+                                                        { kind: 'Field', name: { kind: 'Name', value: 'hasNextPage' } },
+                                                    ],
+                                                },
+                                            },
+                                            { kind: 'Field', name: { kind: 'Name', value: 'totalCount' } },
                                         ],
                                     },
                                 },
