+++ relay/RepoItemFragment.tsx
@@ -1,10 +1,7 @@
 import { graphql, useFragment } from 'react-relay';
-import { Badge, Icon, Link, ListItem, Text } from '@chakra-ui/react';
-import { FontAwesomeIcon } from '@fortawesome/react-fontawesome';
-import { faCodePullRequest } from '@fortawesome/free-solid-svg-icons';
-
-import InternalLink from 'components/InternalLink';
 import { RepoItemFragment_repo$key } from 'relay/__generated__/RepoItemFragment_repo.graphql';
+
+import { RepoItem } from 'components/RepoItem';
 
 export function RepoItemFragment({ repo, hideIfFork }: { repo: RepoItemFragment_repo$key; hideIfFork?: boolean }) {
     const data = useFragment<RepoItemFragment_repo$key>(
@@ -23,27 +20,5 @@ export function RepoItemFragment({ repo, hideIfFork }: { repo: RepoItemFragment_
     );
     if (!data) return null;
 
-    const { name, nameWithOwner, pullRequests, url, isFork } = data;
-    const { totalCount } = pullRequests;
-
-    if (isFork && hideIfFork) return null;
-
-    return (
-        <ListItem alignItems="center" gap="1">
-            <InternalLink prefetch={false} href={'./' + nameWithOwner}>
-                {name}
-            </InternalLink>
-
-            {isFork && <Badge>fork</Badge>}
-            {totalCount > 0 && <span> - {totalCount} PRs</span>}
-            {totalCount > 0 && (
-                <Link href={url} rel="noopener noreferrer nofollow" target="_blank">
-                    - <Text> open in GitHub</Text>
-                    <Icon ml={1}>
-                        <FontAwesomeIcon icon={faCodePullRequest} />
-                    </Icon>
-                </Link>
-            )}
-        </ListItem>
-    );
+    return <RepoItem repo={data} hideIfFork={hideIfFork} />;
 }
